Many updates.
git-svn-id: http://svn.poker.sebster.com/trunk/poker@113 eff34587-6322-df11-911a-000c295dbd0b

+++ poker-core/src/main/java/com/sebster/math/rational/matrix/LUDecomposition.java
@@ -225,7 +225,7 @@ public class LUDecomposition<T extends FieldValue<T> & Comparable<T> & Serializa
 	public double[] getDoublePivot() {
 		double[] vals = new double[m];
 		for (int i = 0; i < m; i++) {
-			vals[i] = (double) piv[i];
+			vals[i] = piv[i];
 		}
 		return vals;
 	}
